+++ dbsupport/transaction.go
@@ -1,4 +1,4 @@
-package update
+package dbsupport
 
 import (
 	"fmt"
@@ -7,9 +7,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-const TenantsUpdateAdvisoryLockID = 4242
-
-func Transaction(db *gorm.DB, lockAndDo func(tx *gorm.DB) error) error {
+func Transaction(db *gorm.DB, lockAndDo LockAndDo) error {
 	var err error
 
 	if db == nil {
@@ -54,17 +52,4 @@ func Transaction(db *gorm.DB, lockAndDo func(tx *gorm.DB) error) error {
 	return nil
 }
 
-type lockAndDo func(tx *gorm.DB) error
-
-func lock(do func(repo Repository) error) lockAndDo {
-	return func(tx *gorm.DB) error {
-		if err := tx.Exec("SET LOCAL lock_timeout = '60s'").Error; err != nil {
-			return errors.Wrap(err, "failed to set lock timeout")
-		}
-		if err := tx.Exec("SELECT pg_advisory_xact_lock($1)", TenantsUpdateAdvisoryLockID).Error; err != nil {
-			return errors.Wrap(err, "failed to acquire lock")
-		}
-
-		return do(NewRepository(tx))
-	}
-}
+type LockAndDo func(tx *gorm.DB) error
